Update: PSをバックグラウンド起動にした

+++ PS1/open_wsl_port.ps1
@@ -1,27 +1,25 @@
-if (!([Security.Principal.WindowsPrincipal][Security.Principal.WindowsIdentity]::GetCurrent()).IsInRole("Administrators")) { Start-Process powershell.exe "-File `"$PSCommandPath`"" -Verb RunAs -Wait; }
-
+if (!([Security.Principal.WindowsPrincipal][Security.Principal.WindowsIdentity]::GetCurrent()).IsInRole("Administrators")) { 
+  Start-Process powershell.exe "-File `"$PSCommandPath`"" -WindowStyle hidden -Verb RunAs -Wait; exit;
+}
 $ip = bash.exe -c "ip r |tail -n1|cut -d ' ' -f9"
 if( ! $ip ){
-  echo "The Script Exited, the ip address of WSL 2 cannot be found";
   exit;
 }
+if (([Security.Principal.WindowsPrincipal][Security.Principal.WindowsIdentity]::GetCurrent()).IsInRole("Administrators")){
+  # All the ports you want to forward separated by comma
+  $ports_a = [Environment]::GetEnvironmentVariable('PORT',"User");
+  $ports=@([Environment]::GetEnvironmentVariable("PORT","User") -Split ",");
 
-# All the ports you want to forward separated by comma
-$ports=@(22,3000,3001,8080);
-$ports_a = $ports -join ",";
-[Environment]::SetEnvironmentVariable('PORT',$ports,"User")
+  if (0 -lt $ports.length){  # Remove Firewall Exception Rules
+    iex "Remove-NetFireWallRule -DisplayName 'WSL 2 Firewall Unlock' ";
 
-# Remove Firewall Exception Rules
-iex "Remove-NetFireWallRule -DisplayName 'WSL 2 Firewall Unlock' ";
+    # Adding Exception Rules for inbound and outbound Rules
+    iex "New-NetFireWallRule -DisplayName 'WSL 2 Firewall Unlock' -Direction Outbound -LocalPort $ports_a -Action Allow -Protocol TCP";
+    iex "New-NetFireWallRule -DisplayName 'WSL 2 Firewall Unlock' -Direction Inbound -LocalPort $ports_a -Action Allow -Protocol TCP";
 
-# Adding Exception Rules for inbound and outbound Rules
-iex "New-NetFireWallRule -DisplayName 'WSL 2 Firewall Unlock' -Direction Outbound -LocalPort $ports_a -Action Allow -Protocol TCP";
-iex "New-NetFireWallRule -DisplayName 'WSL 2 Firewall Unlock' -Direction Inbound -LocalPort $ports_a -Action Allow -Protocol TCP";
-
-for( $i = 0; $i -lt $ports.length; $i++ ){
-  $port = $ports[$i];
-  iex "netsh interface portproxy add v4tov4 listenport=$port listenaddress=* connectport=$port connectaddress=$ip";
+    for( $i = 0; $i -lt $ports.length; $i++ ){
+      $port = $ports[$i];
+      iex "netsh interface portproxy add v4tov4 listenport=$port listenaddress=* connectport=$port connectaddress=$ip";
+    }
+  }
 }
-
-# Show proxies
-iex "netsh interface portproxy show v4tov4";
